+++ frontend/src/index.js
@@ -2,3 +2,6 @@ import './index.css'
 import './components/elements/element-detail-tabs/component';
 import './components/projects/project-view/component';
 import './components/projects/project-view/menu';
+
+import './components/permissions/component';
+import './components/permissions/permissions';